discount test 1/3

+++ exercise.main/Coffee.cs
@@ -8,11 +8,11 @@ namespace exercise.main
     public class Coffee : Item
     {
         public string Code { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public string Name { get; set; }
         public string Variant { get; set; }
 
-        public Coffee(string code, float price, string variant)
+        public Coffee(string code, decimal price, string variant)
         {
             Code = code;
             Price = price;
