+++ exercise.main/Filling.cs
@@ -8,11 +8,11 @@ namespace exercise.main
     public class Filling : Item
     {
         public string Code { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public string Name { get; set; }
         public string Variant { get; set; }
 
-        public Filling(string code, float price, string variant)
+        public Filling(string code, decimal price, string variant)
         {
             Code = code;
             Price = price;
